[MIG] pos_show_clock: Migration to 17.0
