+++ pos_show_clock/static/description/index.html
@@ -411,6 +411,10 @@ If you spotted it first, help us to smash it by providing a detailed and welcome
 <li>Felipe Zago &lt;<a class="reference external" href="mailto:felipe.zago&#64;kmee.com.br">felipe.zago&#64;kmee.com.br</a>&gt;</li>
 </ul>
 </li>
+<li><a class="reference external" href="https://www.heliconia.io">Heliconia Solutions Pvt. Ltd.</a><ul>
+<li>Bhavesh Heliconia</li>
+</ul>
+</li>
 </ul>
 </div>
 <div class="section" id="other-credits">
