+++ pos_show_clock/static/src/css/pos.css
@@ -1,4 +1,3 @@
 .pos .time {
-    color: #ddd;
     font-size: 16px;
 }
